➕ Add web-mode for web development
I have a basic web dev environment set up for React. This commit also
adds `(setq create-lockfiles nil)` because of an error that would
occur when starting React server. The error would complain that a
`.#index.js` file was not found, disabling lockfiles is the solution
to being able to continually run the server.